requirements.txt
fastapi
uvicorn
langchain
langchain-openai
pydantic
openai
python-dotenv
mangum